+++ CHANGELOG.md
@@ -54,6 +54,7 @@ tech changes will usually be stripped from release notes for the public
 -   Token direction UI triggering when other UI is on top of it
 -   Floor detail UI not moving along if side menu is opened
 -   Unlocking shape could sometimes trigger the shape following your mouse
+-   Templates: only using the first character of the provided template name
 
 ## [2022.3.0] - 2022-12-12